addon/pyproject.toml
[project]
name = "ekuraddon"
version = "1.1.6"
description = "A multi-purpose importer for Halo Infinite."
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["fake-bpy-module-latest"]

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = [".venv"]
reportInvalidTypeForm = false
reportIncompatibleMethodOverride = false
reportImplicitOverride = false

[tool.ruff]
line-length = 100

[tool.uv]
dev-dependencies = ["basedpyright<=1.28.3", "pytest<=8.3.5", "ruff<=0.11.2"]
